move comment to the article screen
